change to CustomTextEditor
this allows us to receive changes
